refactor(user): unify user profile routes to /users/[id] structure
- Remove /me/[customId] route structure
- Consolidate all user profiles under /users/[id] with @ prefix for customId
- Move user profile utilities from widgets to entities layer
- Add parseUserId and getUserProfileBasePath helper functions
- Update header navigation to use unified route structure
- Simplify user profile layout and loading components
- Remove redundant UserProfileLayout, UserProfileLoading components
