--- app/[language]/(guest)/(default-layout)/me/[customId]/[[...slug]]/loading.tsx
@@ -1,5 +1,0 @@
-import { UserProfileLoading } from '@/widgets/user';
-
-export default function Loading() {
-  return <UserProfileLoading />;
-}
