--- app/[language]/(guest)/(default-layout)/me/not-found.tsx
@@ -1,5 +1,0 @@
-import { UserProfileNotFound } from '@/widgets/user';
-
-export default function NotFound() {
-  return <UserProfileNotFound />;
-}
