+++ app/[language]/(guest)/(default-layout)/users/[id]/[[...slug]]/loading.tsx
@@ -2,11 +2,11 @@
 
 import { Skeleton } from '@/shared/ui/skeleton';
 
-export const UserProfileLoading = () => {
+export default function Loading() {
   return (
     <div className="space-y-4">
       <Skeleton className="h-48 w-full" />
       <Skeleton className="h-48 w-full" />
     </div>
   );
-};
+}
