--- app/[language]/(guest)/(default-layout)/users/[userId]/[[...slug]]/loading.tsx
@@ -1,5 +1,0 @@
-import { UserProfileLoading } from '@/widgets/user';
-
-export default function Loading() {
-  return <UserProfileLoading />;
-}
