+++ entities/user/index.ts
@@ -1,3 +1,4 @@
 export { getUserData, getUserDataKey } from './api/user-data';
 export type { UserIdentifierType } from './api/user-data';
 export * from './model';
+export * from './lib/user-profile-utils';
