+++ widgets/user/ui/index.ts
@@ -1,6 +1,2 @@
 export * from './user-profile-widget';
 export * from './user-detail-widget';
-export * from './user-profile-layout';
-export * from './user-profile-loading';
-export * from './user-profile-not-found';
-export * from './user-profile-tab-content';
